completed part 4

+++ part4/Blog list/tests/blogApi.test.js
@@ -5,20 +5,38 @@ const testHelper = require("./testHelper");
 const Blog = require("../models/blogModel");
 const api = supertest(app);
 
+let token = ""
+
 beforeEach(async () => {
   await Blog.deleteMany({});
-  await Blog.insertMany(testHelper.initialBlogs);
+  const id = await testHelper.InsertTestBlogs()
+  token = await testHelper.getJwtTokken(id)
 });
 
+
+
 test("Blogs are returned as json", async () => {
+  const token = await testHelper.getJwtTokken()
   await api
     .get("/api/blogs")
+    .set('Authorization', `Bearer ${token}`)
     .expect(200)
     .expect("Content-Type", /application\/json/);
 });
 
+
+test("Blog without token should return 401", async () => {
+  const postData = {
+    title: "sampledasdasd",
+    author: "xd",
+    url: "example",
+    likes: "69",
+  };
+  await api.post("/api/blogs").send(postData).expect(401);
+})
+
 test("Blogs json should have a unique identifier", async () => {
-  (await api.get("/api/blogs")).body.forEach((doc) => {
+  (await api.get("/api/blogs").set('Authorization', `Bearer ${token}`)).body.forEach((doc) => {
     expect(doc.id).toBeDefined();
   });
 });
@@ -33,11 +51,12 @@ test("testing blog post method", async () => {
     likes: "69",
   };
 
-  await api.post("/api/blogs").send(postData).expect(201);
+  await api.post("/api/blogs").set('Authorization', `Bearer ${token}`).send(postData).expect(201);
 
-  const LenAfterInsert = (await api.get("/api/blogs")).body.length;
+  const LenAfterInsert = (await api.get("/api/blogs").set('Authorization', `Bearer ${token}`)).body.length;
 
   expect(LenAfterInsert).toBe(LenBeforeInsert + 1);
+
 });
 
 test("when like prop is not supplied it will default to zero", async () => {
@@ -47,7 +66,7 @@ test("when like prop is not supplied it will default to zero", async () => {
     url: "example",
   };
 
-  const result = await api.post("/api/blogs").send(postData).expect(201);
+  const result = await api.post("/api/blogs").set('Authorization', `Bearer ${token}`).send(postData).expect(201);
   expect(result.body.likes).toBe(0);
 });
 
@@ -60,22 +79,22 @@ test("Expect 422 when title or url missing in post data", async () => {
     title: "sampledasdasd",
     author: "xd",
   };
-  await api.post("/api/blogs").send(postData).expect(422);
-  await api.post("/api/blogs").send(_postData).expect(422);
+  await api.post("/api/blogs").set('Authorization', `Bearer ${token}`).send(postData).expect(422);
+  await api.post("/api/blogs").set('Authorization', `Bearer ${token}`).send(_postData).expect(422);
 });
 
 describe("deletion of a blog", () => {
   test("succeeds with status code 204 if id is valid", async () => {
     const blogsAtStart = await testHelper.blogsInDb();
     const blogToDelete = blogsAtStart[0];
-    await api.delete(`/api/blogs/${blogToDelete.id}`).expect(204);
+    await api.delete(`/api/blogs/${blogToDelete.id}`).set('Authorization', `Bearer ${token}`).expect(204);
     const blogsAtEnd = await testHelper.blogsInDb();
     expect(blogsAtEnd).toHaveLength(testHelper.initialBlogs.length - 1);
   });
 
   test("404 if id not exist in db", async () => {
     const id = await testHelper.nonExistingId();
-    await api.delete(`/api/blogs/${id}`).expect(404);
+    await api.delete(`/api/blogs/${id}`).set('Authorization', `Bearer ${token}`).expect(404);
   });
 });
 
@@ -85,7 +104,7 @@ describe("updation of a blog", () => {
     const blogToUpdate = blogsAtStart[0];
     blogToUpdate.title = "this part is updated";
     await api
-      .put(`/api/blogs/${blogToUpdate.id}`)
+      .put(`/api/blogs/${blogToUpdate.id}`).set('Authorization', `Bearer ${token}`)
       .send(blogToUpdate)
       .expect(200);
   });
@@ -97,7 +116,7 @@ describe("updation of a blog", () => {
       author: "xd",
       url: "example",
     };
-    await api.put(`/api/blogs/${id}`).send(postData).expect(404);
+    await api.put(`/api/blogs/${id}`).set('Authorization', `Bearer ${token}`).send(postData).expect(404);
   });
 });
